requirements.txt
azure-functions
azure-storage-file-share
azure-keyvault-secrets
azure-identity